refactor: Abstract out the side panel logic
